Add separate api structure
